refactoring #5 - RecentBooksCollectionView를 SearchResultCollectionViewHeader에서 분리하기
